Added win/lose conditions
If the timer ends with the player not in the escape point, the player is greeted with a game over screen with the score, with a button back to the menu.

If the players get to the escape point before the timer runs out, the player is greeted with a win screen with the score, with a button back to the menu.